+++ ttkeditor/app/about.py
@@ -25,18 +25,18 @@
 
 from TermTk.TTkCore.color import TTkColor
 from TermTk import TTkWindow
-from .config import TTkEditConfig
+from .config import TTKEditorConfig
 
-class TTkEditAbout(TTkWindow):
+class TTKEditorAbout(TTkWindow):
     ttkeditor = [
-        "__________________    _______                      ",
-        "\_______________ /   / / ____)  ┌─┐─┐─┐            ",
-        "    /\    /\   | |__/ /| |__  __| |─┘ └─┐ __   ___ ",
-        "   |  |  |  |  |  _  / |  __|/ _  |─┐  ┌┘/ _ \|  _|",
-        "   |  |  |  |  | | \ \ | |__( (_| | |  \( (_| ) |  ",
-        "   |  |  |  |  └─┘  \_)|_____)____|_|\___)___/|_|  ",
-        "   |  |  |  |                                      ",
-        "   └──┘  └──┘                                      ",]
+        r"__________________    _______                      ",
+        r"\_______________ /   / / ____)  ┌─┐─┐─┐            ",
+        r"    /\    /\   | |__/ /| |__  __| |─┘ └─┐ __   ___ ",
+        r"   |  |  |  |  |  _  / |  __|/ _  |─┐  ┌┘/ _ \|  _|",
+        r"   |  |  |  |  | | \ \ | |__( (_| | |  \( (_| ) |  ",
+        r"   |  |  |  |  └─┘  \_)|_____)____|_|\___)___/|_|  ",
+        r"   |  |  |  |                                      ",
+        r"   └──┘  └──┘                                      ",]
 
     def __init__(self, *args, **kwargs):
         TTkWindow.__init__(self, *args, **kwargs)
@@ -46,11 +46,11 @@ class TTkEditAbout(TTkWindow):
 
     def paintEvent(self, canvas):
         c = [0xFF,0xFF,0xFF]
-        for y, line in enumerate(TTkEditAbout.ttkeditor):
+        for y, line in enumerate(TTKEditorAbout.ttkeditor):
             canvas.drawText(pos=(3,3+y),text=line, color=TTkColor.fg(f'#{c[0]:02X}{c[1]:02X}{c[2]:02X}'))
             c[2]-=0x18
             c[0]-=0x08
-        canvas.drawText(pos=(35,10),text=f"  Version: {TTkEditConfig.version}", color=TTkColor.fg('#AAAAFF'))
+        canvas.drawText(pos=(35,10),text=f"  Version: {TTKEditorConfig.version}", color=TTkColor.fg('#AAAAFF'))
         canvas.drawText(pos=( 6,12),text=f"https://github.com/guillermomolina/ttkeditor", color=TTkColor.fg('#44FFFF'))
 
         TTkWindow.paintEvent(self, canvas)